% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_KL.R
\name{calculateKL_abs}
\alias{calculateKL_abs}
\title{Calculate a modified KL score that uses absolute value of the ratios
It will detect changes in the relative proportion of enzymes connected to a metabolite}
\usage{
calculateKL_abs(x, y)
}
\arguments{
\item{x}{expression vector for the reference/baseline condition, in FPKM, or TPM value - normalized for depth and gene length.}

\item{y}{expression vector for the perturbed/treated condition, in FPKM, or TPM value - normalized for depth and gene length.}
}
\value{
KL value denoting the difference between the distribution of values in x and y
}
\description{
Calculate a modified KL score that uses absolute value of the ratios
It will detect changes in the relative proportion of enzymes connected to a metabolite
}
