% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_KL.R
\name{calculateKL_Dirichlet}
\alias{calculateKL_Dirichlet}
\title{Calculate a modified KL score that uses absolute value of the ratios
It will detect cases where the relative proportion did not change, but the total amount of expressed enzymes changed.}
\usage{
calculateKL_Dirichlet(x, y)
}
\arguments{
\item{x}{expression vector for the reference/baseline condition, in FPKM, or TPM value - normalized for depth and gene length.}

\item{y}{expression vector for the perturbed/treated condition, in FPKM, or TPM value - normalized for depth and gene length.}
}
\value{
KL value denoting the difference between the distribution of values in x and y
}
\description{
Calculate a modified KL score that uses absolute value of the ratios
It will detect cases where the relative proportion did not change, but the total amount of expressed enzymes changed.
}
