% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPerturbMet_noPerm.R
\name{findPerturbMet_noPerm}
\alias{findPerturbMet_noPerm}
\title{Identify perturbed metabolites based on changes in the relative expression of enzymes around a metabolite
This function does NOT perform metabolic network topology permutation. This is intended to be used when performing sample label permutation.}
\usage{
findPerturbMet_noPerm(expression_data, mets2genes, expressed_threshold)
}
\arguments{
\item{expression_data}{data frame of the input expression data.First column must be gene symbol; second column is the expression level in reference condition; third comun is the expression level in treated/perturbed condition; expression in TPM or FPKM, not log transformed.}

\item{mets2genes}{data frame that describe gene-metabolite connections. It includes two columns: "met", which lists metabolite symbols; and "genes" which include gene symbols for genes that produce or consume the corresponding metabolite}

\item{expressed_threshold}{a value above which a gene is considered expressed}
}
\value{
a list that has two components. 1. perturbed_mets: list of perturbed metabolites sorted by p-value; 2. met_gene_pairs: metabolite-gene pairs after filtering down to expressed genes and their connected metabolites;
}
\description{
Identify perturbed metabolites based on changes in the relative expression of enzymes around a metabolite
This function does NOT perform metabolic network topology permutation. This is intended to be used when performing sample label permutation.
}
