% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pct_change.R
\name{plot_pct_change}
\alias{plot_pct_change}
\title{Make a bar plot comparing the percentage changes of all enzymes connected to a metabolite between two conditions.}
\usage{
plot_pct_change(met_gene_exprs, met2look, legend_coord, n2plot, ylimit)
}
\arguments{
\item{met_gene_exprs}{a dataframe that includes metabolite-gene connections and gene expression values.}

\item{met2look}{which metabolite to plot, using metabolite symbols}

\item{legend_coord}{where to place the legend. c(0,0) is bottom left; c(1,1) is top right.}

\item{n2plot}{how many genes to plot}

\item{ylimit}{y axis limit}
}
\value{
Return a plot of bar plots
}
\description{
Make a bar plot comparing the percentage changes of all enzymes connected to a metabolite between two conditions.
}
